pyproject.toml
[project]
name = "agenticsports"
version = "0.1.0"
description = "AgenticSports - Autonomous AI Sports Coach"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0",
    "fitdecode>=0.10",
    "rich>=14.0",
    "watchdog>=6.0",
    "python-dotenv>=1.0",
    "numpy>=2.4.2",
    "bm25s>=0.2.14",
    "litellm>=1.60",
    "pydantic-settings>=2.13.0",
    "supabase>=2.28.0",
    # -- FastAPI / Server ------------------------------------------------------
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.0",
    "python-multipart>=0.0.9",
    # -- Auth / Security -------------------------------------------------------
    "PyJWT>=2.8",
    "cryptography>=41",
    # -- Rate limiting / caching -----------------------------------------------
    "slowapi>=0.1.9",
    "redis>=5.0",
    # -- Calc engine -----------------------------------------------------------
    "evalidate>=2.0",
    # -- HTTP client -----------------------------------------------------------
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that call external APIs (Gemini)",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "fakeredis>=2.21",
]
